Revert "Move to opencv apps"
This reverts commit 6748a24b622a4513617cbcb4f40b93d9e730ff74.
