+++ apps/CMakeLists.txt
@@ -60,4 +60,3 @@ ocv_add_app(visualisation)
 ocv_add_app(interactive-calibration)
 ocv_add_app(version)
 ocv_add_app(model-diagnostics)
-ocv_add_app(pipeline-modeling-tool)
