+++ modules/gapi/samples/pipeline_modeling_tool/configs/config_template.yml
