+++ modules/gapi/samples/pipeline_modeling_tool/dummy_source.hpp
